feat(CIT-341): Visualización de historial de turnos
