+++ modules/turnos/controller/historialCitasController/historialCitasController.ts
@@ -1,31 +1,67 @@
 import { search as searchPrestaciones } from '../../../../modules/rup/controllers/prestacion';
-import { getHistorialPaciente, getLiberadosPaciente } from '../../controller/turnosController';
-import moment = require('moment');
+import { getLiberadosPaciente } from '../../controller/turnosController';
+import { getHistorialPaciente } from '../../../../core-v2/mpi/paciente/paciente.controller';
 import { PacienteCtr } from '../../../../core-v2/mpi/paciente/paciente.routes';
+import moment = require('moment');
+import * as mongoose from 'mongoose';
 
 export async function getHistorial(req) {
+    let historial = [];
+    const turnosPaciente = [];
 
-    const turnos = getHistorialPaciente(req);
-    const liberados = getLiberadosPaciente(req);
-    const fueraAgendas = getHistorialFueraAgendas(req.query.pacienteId);
-    const result = await Promise.all([turnos, liberados, fueraAgendas]);
-    const historial = [...result[0], ...result[1], ...result[2]];
+    const estado = req.query.estado;
 
+    const acciones = [
+        { estados: [undefined, 'asignado', 'suspendido'], fn: getHistorialPaciente },
+        { estados: [undefined, 'fuera de agenda'], fn: getHistorialFueraAgendas },
+        { estados: [undefined, 'liberado'], fn: getLiberadosPaciente }
+    ];
+
+    for (const { estados, fn } of acciones) {
+        if (estados.includes(estado)) {
+            turnosPaciente.push(fn(req));
+        }
+    }
+
+    const result = await Promise.all(turnosPaciente);
+
+    if (result.length === 1) {
+        historial = [...result[0]];
+    } else {
+        historial = [...result[0], ...result[1], ...result[2]];
+    }
     return historial;
 }
 
-async function getHistorialFueraAgendas(pacienteId) {
-    const paciente = await PacienteCtr.findById(pacienteId);
+async function getHistorialFueraAgendas(req) {
+    const paciente = await PacienteCtr.findById(req.query.pacienteId);
 
     const paramsFueraAgenda = {
         'paciente.id': { $in: paciente.vinculos },
         'estadoActual.tipo': 'validada',
         inicio: 'fuera-agenda',
         'solicitud.ambitoOrigen': 'ambulatorio',
-        'ejecucion.fecha': { $gt: moment().subtract(1, 'year').startOf('day').toDate() },
         'solicitud.turno': null
     };
 
+    if (req.query.desde && req.query.hasta) {
+        paramsFueraAgenda['ejecucion.fecha'] = {
+            $gte: moment(req.query.desde).startOf('day').toDate(),
+            $lte: moment(req.query.hasta).endOf('day').toDate()
+        };
+    } else if (req.query.desde) {
+        paramsFueraAgenda['ejecucion.fecha'] = {
+            $gte: moment(req.query.desde).startOf('day').toDate(),
+        };
+    } else if (req.query.hasta) {
+        paramsFueraAgenda['ejecucion.fecha'] = {
+            $lte: moment(req.query.hasta).endOf('day').toDate()
+        };
+    }
+
+    if (req.query.prestacion) {
+        paramsFueraAgenda['solicitud.tipoPrestacion.id'] = mongoose.Types.ObjectId(req.query.prestacion);
+    }
     const fueraAgendas = await searchPrestaciones(paramsFueraAgenda);
 
     const historialFueraAgendas = fueraAgendas.map((elem: any) => ({
