+++ modules/turnos/routes/turno.ts
@@ -33,7 +33,9 @@ router.get('/turno/:id*?', async (req, res, next) => {
 });
 router.get('/historial', async (req, res, next) => {
     try {
-        res.json(await getHistorial(req));
+        let turno = await getHistorial(req);
+        turno = turno.splice(req.query.skip, req.query.limit);
+        res.json(turno);
     } catch (err) {
         return next(err);
     }
